refactor(desktop): extract login-shell discovery into submodule
discovery.rs sits far over the file-size ratchet ceiling for
pre-existing reasons, so it may not grow. Move the login-shell PATH
probe, its cache, and the nvm default-bin helpers into
discovery/login_shell.rs verbatim (visibility + import adjustments
only) to bring discovery.rs back under the ceiling. Pure move: no
behavior or signature changes.
